WIP - Add support for `i32` sample format type.
This allows for avoiding a currently-necessary conversion in the ASIO
host implementation. It is necessary for the ASIO host to support i32
sample formats as the majority of ASIO drivers only seem to support this
format, including ASIO4All.

Originally, this branch was started to rule out the possibility of the
sample format conversion being the cause of perceived distortion during
playback using ASIO4ALL. The distortion remains, so the cause must
indeed be something else.

Only the WASAPI and ASIO hosts have been updated for the introduction of
`i32` support so far.